make_ext4fs: add command-line -u option for UUID
As the ability to easily determine or specify the UUID of the
file-system eases working with some bootloaders, this patch adds
printing of the UUID upon file system creation with the other
information and allows for the UUID to be specified as a command-line
parameter.

Whether specified or automatically generated, the UUID is now printed
with the Label and other creation information.

The parsing of command-line supplied UUIDs is relatively forgiving,
allowing dash-separated values or not.

The default UUID generation remains unchanged as the hash of the
file-system label.

Signed-off-by: Eric Herman <eric@freesa.org>